Fixes duplicated IdentityProvider definition
