+++ ts/models/IdentityProvider.ts
@@ -1,0 +1,15 @@
+/**
+ * Describes a SPID Identity Provider
+ */
+
+export type IdentityProvider = {
+  id: string;
+  logo: any;
+  name: string;
+  entityID: string;
+  profileUrl: string;
+};
+
+export function isDemoIdp(idp: IdentityProvider): boolean {
+  return idp.id === "demo";
+}
