+++ ts/utils/api.ts
@@ -69,18 +69,3 @@ export async function setUserProfile(
     // @see https://www.pivotaltracker.com/story/show/154661120
   }
 }
-
-/**
- * Describes a SPID Identity Provider
- */
-export type IdentityProvider = {
-  id: string;
-  logo: any;
-  name: string;
-  entityID: string;
-  profileUrl: string;
-};
-
-export function isDemoIdp(idp: IdentityProvider): boolean {
-  return idp.id === "demo";
-}
